Switched to puppeteer for dom scraping, changed some styling
Since musescore changed the way they are hosting the pages of their documents, I switched over to puppeteer for a failsafe method of retrieving the links. Whether musescore did this on purpose to prevent link grabbers from getting the links to their documents or whether it was just coincidental, only the first image seems to be hosted on their servers, with the rest bein spread out on AWS, ulimate-guitar and others (from what I can gather). The only way to get the correct links now is to manually scrolling the site, which I am doing with puppeteer now. This should hold for a while. ;)